some improvements and bug fixing

+++ zt.py
@@ -16,13 +16,13 @@ import select
 Preset for some globale variable, they may be changed via
 commandline argumemts
 """
-conf_file='/etc/zt.conf'
-pid_file='/tmp/zt.pid'
+conf_file='/etc/zt_nss/zt.conf'
+pid_file='/var/run/zt.pid'
 
 """
 This must be set according to the define within nss_zt.c
 """
-sock_file='/tmp/zt.sock' 
+sock_file='/var/run/zt.sock' 
 
 """
 Srorage from data collected from zerotier
@@ -39,17 +39,20 @@ def read_config(conf_file):
 	return url, headers
 
 
-def searchName(list, name):
+def searchName(list, name,noact):
 	str = name.split('.')
 	num=len(str)
 	if num > 1 :
 		nm = str[num-2]+'.'+str[num-1]
 		for sys in list:
-			#if sys[0] == nm:
-			if sys[0] == nm and sys[2]:
-				return sys[1]
+			if noact == True:
+				if sys[0] == nm:
+					return sys[1]
+			else:
+				if sys[0] == nm and sys[2]:
+					return sys[1]
 
-def searchIp(list, ip):
+def searchIp(list, ip, noact):
 	for sys in list:
 		if sys[1] == ip:
 			return sys[0]
@@ -73,14 +76,14 @@ def requestHosts(url, headers, timeout, debug):
 				list.append(elem)
 		time.sleep(timeout)
 
-def processMessage(sock,debug):
+def processMessage(sock,noact, debug):
 	connection, client_address = sock.accept()
 	datagram = connection.recv(1024)
 	if datagram:
 		print(datagram)
 	what, nm = datagram.decode('utf-8').split('|')
 	if what  == 'N':
-		ip = searchIp(list,nm)
+		ip = searchIp(list,nm, noact)
 	else:
 		if debug:
 			print('Look for name', nm)
@@ -93,7 +96,7 @@ def processMessage(sock,debug):
 			if debug:
 				print('wrong to domain')
 			return
-		ip = searchName(list,nm)
+		ip = searchName(list,nm, noact)
 		if debug:
 			print(ip)
 	# send answer
@@ -105,10 +108,8 @@ def processMessage(sock,debug):
 		connection.sendall(b'')
 		connection.shutdown(socket.SHUT_RDWR)
 		connection.close()
-	
 
-def zt_server(url,headers, timeout, bind, port, debug):
-	#requestHosts(url,headers, timeout, debug)
+def zt_server(url,headers,noact,timeout,bind,port,debug):
 	t = threading.Thread(target=requestHosts, args=(url, headers, timeout, debug,))
 	t.daemon = True
 	t.start()	
@@ -133,13 +134,14 @@ def zt_server(url,headers, timeout, bind, port, debug):
 		except:
 			# Wrong Address or port?
 			print("Can't connect to",bind,"port",port)
-			os.remove(sock_file)
+			if os.path.exists(sock_file):
+				os.remove(sock_file)
 			sys.exit(2);
 	while True:
 		try:
 			r,w,e = select.select(sockets,[],[],50000)
 			for s in r:
-				processMessage(s,debug)
+				processMessage(s,noact,debug)
 		except TimeoutError:
 			if debug:
 				print('TimeoutError')
@@ -149,11 +151,13 @@ def zt_server(url,headers, timeout, bind, port, debug):
 			if debug:
 				print('KeyboardInterrupt')
 			sock.close()
-			os.remove(sock_file)
+			if os.path.exists(sock_file):
+				os.remove(sock_file)
 			break;
 		except SystemExit:
 			sock.close();
-			os.remove(sock_file)
+			if os.path.exists(sock_file):
+				os.remove(sock_file)
 			return
 		except:
 			reason=str(sys.exc_info()[1])
@@ -165,7 +169,7 @@ def zt_server(url,headers, timeout, bind, port, debug):
 			else:
 				return
 
-def run(pidf,url,headers,timeout,bind,port,debug):
+def run(pidf,url,headers,noact,timeout,bind,port,debug):
 	if debug:
 		print('run()')
 		dbg_file='/tmp/zt.txt'
@@ -181,32 +185,37 @@ def run(pidf,url,headers,timeout,bind,port,debug):
 			stdout= open(dbg_file, 'w+'),
 			stderr= open(err_file, 'w+'),
         ) as context:
-		zt_server(url,headers, timeout, bind, port, debug)
+		zt_server(url,headers,noact,timeout,bind,port,debug)
 
 if __name__ == "__main__":
 	parser = argparse.ArgumentParser(description="Zerotier DNS server")
 	parser.add_argument('-p', '--pid-file', default=pid_file)
-	#parser.add_argument('-s', '--socket-file', default=sock_file)
 	parser.add_argument('-c', '--config-file', default=conf_file)
 	parser.add_argument('-t', '--timeout',type=int, default=600)
+	parser.add_argument('-a', '--all', action='store_true')
 	parser.add_argument('-f', '--foreground', action='store_true')
 	parser.add_argument('-d', '--debug', action='store_true')
 	parser.add_argument('-b', '--bind', default=None)
 	parser.add_argument('-P', '--port',type=int, default=9999)
 	parser.add_argument('-v', '--version', action='store_true')
 	args = parser.parse_args()
-	timeout=args.timeout
+	
 	if args.version:
-		print('version: 20190002')
+		print('version: 20190003')
+		print(conf.port)
 		sys.exit(0)
 	conf_file=args.config_file
 	if os.path.exists(sock_file):
 		print('Error: Socket file exist')
 		sys.exit(1)
 	url, headers = read_config(conf_file)
+	noact=args.all;
+	conf.url=url
+	conf.headers=headers
 	if args.foreground == True:
-		zt_server(url, headers,args.timeout,args.bind,args.port,args.debug)
+		zt_server(url,headers,noact,args.timeout,args.bind,args.port,args.debug)
 	else:
-		run(args.pid_file, url, headers, args.timeout,args.bind,args.port,args.debug)
-		#os.remove(sock_file)
+		run(args.pid_file, url,headers,noact,args.timeout,args.bind,args.port,args.debug)
+	if os.path.exists(sock_file):
+		os.remove(sock_file)
 	sys.exit(0)
